fix(capture): sample changeCount baseline at the first Cmd+C
The baseline changeCount was read in handleDoubleCopy() at the moment the
double-press was detected. NSEvent.addGlobalMonitorForEvents is a passive
observer that fires *after* the foreground app handles the keyDown, so an
app that copies synchronously on Cmd+C (Chromium/Electron, Java, some
custom NSResponder text views) has already bumped changeCount by then. The
baseline then equals the post-copy value, the poll loop's
`changeCount > baseline` predicate never holds within the budget, and the
user sees "Nie udało się pobrać zaznaczenia" despite a correct clipboard.

Sample the baseline at the *first* Cmd+C instead: GlobalHotkeyMonitor now
stashes the changeCount on the first press (registerPress) and passes it
through onDoubleCopy(baseline:), so it always precedes the second copy
regardless of how synchronously the app copies. The changeCount source is
an injected provider (wired from the reader in AppDelegate) to keep the
monitor decoupled and testable. stop() clears the pending baseline along
with the detector window.

Changes the HotkeyMonitor seam (onDoubleCopy now carries the baseline).
Adds GlobalHotkeyMonitorTests for the first-press/rebaseline logic.

+++ Tests/Support/CoordinatorFakes.swift
@@ -75,7 +75,7 @@ final class FakePopup: TranslationPopupPresenting {
 
 @MainActor
 final class FakeHotkeyMonitor: HotkeyMonitor {
-    var onDoubleCopy: (@MainActor () -> Void)?
+    var onDoubleCopy: (@MainActor (Int) -> Void)?
     private(set) var stopCount = 0
     func start() throws {}
     func stop() { stopCount += 1 }
